[src] Adjust format

+++ src/poller.cc
@@ -125,7 +125,7 @@ void Poller::RemoveEventer(Eventer* eventer) {
 
 void Poller::GetActiveEventer(int event_amount,
                               EventerList* active_eventers) const {
-  for (int i = 0; i < event_amount; ++i) {
+  for (size_t i = 0; i < event_amount; ++i) {
     auto eventer = static_cast<Eventer*>(poll_events_[i].data.ptr);
     eventer->ReceiveEvents(poll_events_[i].events);
     active_eventers->emplace_back(eventer);
@@ -244,7 +244,8 @@ void Poller::RemoveEventer(Eventer* eventer) {
   poll_events_.pop_back();
 }
 
-void Poller::GetActiveEventer(int event_amount, const PollEventList active_events,
+void Poller::GetActiveEventer(int event_amount,
+                              const PollEventList active_events,
                               EventerList* active_eventers) const {
   LockGuard lock_guard(event_lock_);
   for (auto itr = active_events.cbegin();
